Moved fido2 popup to vault folder on the browser
