+++ apps/browser/src/vault/popup/components/fido2/fido2.component.html
